requirements.txt
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bleach==1.5.0
cycler==0.10.0
html5lib==0.9999999
Keras==2.0.8
krakenex==0.1.4
Markdown==2.6.8
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.3
progress==1.3
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
scikit-learn==0.19.0
scipy==0.19.1
selenium==3.5.0
six==1.10.0
sklearn==0.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
Werkzeug==0.12.2
